+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+* Propagation of flattening attributes in some cases.
+
 ## [0.27.1]
 
 ## Added